// backend/src/types/providerNames.ts
export enum ProviderNames {
  GODADDY = 'godaddy',
  NAMECHEAP = 'namecheap',
  SQUARESPACE = 'squarespace',
  HOSTINGER = 'hostinger',
  NETWORKSOLUTIONS = 'networksolutions',
  NAMECOM = 'namecom',
  PORKBUN = 'porkbun',
  IONOS = 'ionos',
  HOVER = 'hover',
  DYNADOT = 'dynadot',
  NAMESILO = 'namesilo',
  SPACESHIP = 'spaceship'
}